Fixed threading (thread safe TT) and parsing 'Threads'

+++ src/State.h
@@ -75,7 +75,7 @@ namespace engine
 		std::optional<Position> en_passant_target_square{std::nullopt};
 		std::uint64_t zobrist_hash;
 		std::vector<std::uint64_t> repetition_history{};
-		double evaluation{0};
+		int evaluation{0};
 
 		void make(const Move& move) noexcept;
 		void unmove() noexcept;
@@ -104,7 +104,7 @@ namespace engine
 			Castling_rights_map<bool> black_castling_rights;
 			std::uint64_t previous_zobrist_hash;
 			unsigned half_move_clock;
-			double evaluation;
+			int evaluation;
 
 			constexpr bool operator==(const State_delta& state_delta) const = default;
 		};
